PAAC-668: Fixed unrelated messages to all input fields when user enter 5000000000 (Bigger value).

+++ app/form/Period1TriggerDCForm.scala
@@ -31,7 +31,7 @@ trait Period1TriggerDCForm extends TriggerDCFormFactory {
       Form[Period1TriggerDCModel](mapping(poundsPeriod1TEDC)(toModel)(Period1TriggerDCModel.unapply))
   }
 
-  protected val toModel: (Int) => Period1TriggerDCModel = (i: Int) =>
+  protected val toModel: (Long) => Period1TriggerDCModel = (i: Long) =>
     Period1TriggerDCModel(BigDecimal(i))
 }
